refactor(admin): extract Iframe/Redirect sections from AppFormEditor (#1781)
Pulls the "Only for iframe apps" and "Only for redirect apps" cards out of
the 2,902-line AppFormEditor.jsx into their own components
(IframeConfigSection.jsx, RedirectConfigSection.jsx), following the same
incremental-extraction pattern already used for the Upload, Variables,
Starter Prompts, Tools/MCP, Web Search/iAssistant, and Transcription
sections against this issue. Pure extraction, no behavior change.

AppFormEditor.jsx shrinks from 2,902 to 2,699 lines.
