+++ client/src/features/admin/components/AppFormEditor.jsx
@@ -7,6 +7,8 @@ import SkillsSelector from '../../../shared/components/SkillsSelector';
 import WorkflowsSelector from '../../../shared/components/WorkflowsSelector';
 import SourcePicker from './SourcePicker';
 import ResourceSelector from './ResourceSelector';
+import IframeConfigSection from './app-form/IframeConfigSection';
+import RedirectConfigSection from './app-form/RedirectConfigSection';
 import Icon from '../../../shared/components/Icon';
 import IconPicker from '../../../shared/components/IconPicker';
 import MimeTypeSelector from './MimeTypeSelector';
@@ -728,217 +730,12 @@ function AppFormEditor({
 
         {/* Iframe Configuration - Only for iframe apps */}
         {app.type === 'iframe' && (
-          <div className="bg-white dark:bg-gray-800 shadow px-4 py-5 sm:rounded-lg sm:p-6">
-            <div className="md:grid md:grid-cols-3 md:gap-6">
-              <div className="md:col-span-1">
-                <h3 className="text-lg font-medium leading-6 text-gray-900 dark:text-gray-100">
-                  {t('admin.apps.edit.iframeConfig', 'Iframe Configuration')}
-                </h3>
-                <p className="mt-1 text-sm text-gray-500 dark:text-gray-400">
-                  {t('admin.apps.edit.iframeConfigDesc', 'Configure the embedded iframe content')}
-                </p>
-              </div>
-              <div className="mt-5 md:col-span-2 md:mt-0">
-                <div className="space-y-4">
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 dark:text-gray-300">
-                      {t('admin.apps.edit.iframeUrl', 'Iframe URL')}
-                      <span className="text-red-500 ml-1">*</span>
-                    </label>
-                    <input
-                      type="url"
-                      value={app.iframeConfig?.url || ''}
-                      onChange={e =>
-                        handleInputChange('iframeConfig', {
-                          ...app.iframeConfig,
-                          url: e.target.value
-                        })
-                      }
-                      placeholder="https://example.com"
-                      className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
-                    />
-                  </div>
-                  <div className="flex items-center">
-                    <input
-                      type="checkbox"
-                      checked={app.iframeConfig?.allowFullscreen !== false}
-                      onChange={e =>
-                        handleInputChange('iframeConfig', {
-                          ...app.iframeConfig,
-                          allowFullscreen: e.target.checked
-                        })
-                      }
-                      className="h-4 w-4 text-indigo-600 focus:ring-indigo-500 border-gray-300 dark:border-gray-600 rounded"
-                    />
-                    <label className="ml-2 block text-sm text-gray-900 dark:text-gray-100">
-                      {t('admin.apps.edit.allowFullscreen', 'Allow Fullscreen')}
-                    </label>
-                  </div>
-
-                  {/* Sandbox Configuration */}
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 mb-2">
-                      {t('admin.apps.edit.sandboxPermissions', 'Sandbox Permissions')}
-                    </label>
-                    <p className="text-xs text-gray-500 mb-3">
-                      {t(
-                        'admin.apps.edit.sandboxPermissionsDesc',
-                        'Control what the embedded content is allowed to do. Uncheck options to restrict permissions for better security.'
-                      )}
-                    </p>
-                    <div className="space-y-2 pl-4">
-                      {[
-                        {
-                          value: 'allow-scripts',
-                          label: t(
-                            'admin.apps.edit.allowScripts',
-                            'Allow Scripts (required for most interactive content)'
-                          )
-                        },
-                        {
-                          value: 'allow-same-origin',
-                          label: t(
-                            'admin.apps.edit.allowSameOrigin',
-                            'Allow Same Origin (enables localStorage, cookies)'
-                          )
-                        },
-                        {
-                          value: 'allow-forms',
-                          label: t(
-                            'admin.apps.edit.allowForms',
-                            'Allow Forms (enable form submission)'
-                          )
-                        },
-                        {
-                          value: 'allow-popups',
-                          label: t('admin.apps.edit.allowPopups', 'Allow Popups')
-                        },
-                        {
-                          value: 'allow-modals',
-                          label: t(
-                            'admin.apps.edit.allowModals',
-                            'Allow Modals (alert, confirm, etc.)'
-                          )
-                        },
-                        {
-                          value: 'allow-top-navigation',
-                          label: t(
-                            'admin.apps.edit.allowTopNavigation',
-                            'Allow Top Navigation (navigate parent window)'
-                          )
-                        }
-                      ].map(permission => {
-                        const currentSandbox =
-                          app.iframeConfig?.sandbox ||
-                          (app.iframeConfig
-                            ? []
-                            : ['allow-scripts', 'allow-same-origin', 'allow-forms']);
-                        const isChecked = currentSandbox.includes(permission.value);
-
-                        return (
-                          <div key={permission.value} className="flex items-center">
-                            <input
-                              type="checkbox"
-                              checked={isChecked}
-                              onChange={e => {
-                                const currentSandbox = app.iframeConfig?.sandbox || [
-                                  'allow-scripts',
-                                  'allow-same-origin',
-                                  'allow-forms'
-                                ];
-                                const newSandbox = e.target.checked
-                                  ? [...currentSandbox, permission.value]
-                                  : currentSandbox.filter(p => p !== permission.value);
-
-                                handleInputChange('iframeConfig', {
-                                  ...app.iframeConfig,
-                                  sandbox: newSandbox
-                                });
-                              }}
-                              className="h-4 w-4 text-indigo-600 focus:ring-indigo-500 border-gray-300 dark:border-gray-600 rounded"
-                            />
-                            <label className="ml-2 block text-sm text-gray-700">
-                              {permission.label}
-                            </label>
-                          </div>
-                        );
-                      })}
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
+          <IframeConfigSection app={app} onChange={handleInputChange} t={t} />
         )}
 
         {/* Redirect Configuration - Only for redirect apps */}
         {app.type === 'redirect' && (
-          <div className="bg-white dark:bg-gray-800 shadow px-4 py-5 sm:rounded-lg sm:p-6">
-            <div className="md:grid md:grid-cols-3 md:gap-6">
-              <div className="md:col-span-1">
-                <h3 className="text-lg font-medium leading-6 text-gray-900 dark:text-gray-100">
-                  {t('admin.apps.edit.redirectConfig', 'Redirect Configuration')}
-                </h3>
-                <p className="mt-1 text-sm text-gray-500 dark:text-gray-400">
-                  {t('admin.apps.edit.redirectConfigDesc', 'Configure the redirect behavior')}
-                </p>
-              </div>
-              <div className="mt-5 md:col-span-2 md:mt-0">
-                <div className="space-y-4">
-                  <div>
-                    <label className="block text-sm font-medium text-gray-700 dark:text-gray-300">
-                      {t('admin.apps.edit.redirectUrl', 'Redirect URL')}
-                      <span className="text-red-500 ml-1">*</span>
-                    </label>
-                    <input
-                      type="url"
-                      value={app.redirectConfig?.url || ''}
-                      onChange={e =>
-                        handleInputChange('redirectConfig', {
-                          ...app.redirectConfig,
-                          url: e.target.value
-                        })
-                      }
-                      placeholder="https://example.com"
-                      className="mt-1 block w-full rounded-md border-gray-300 shadow-sm focus:border-indigo-500 focus:ring-indigo-500 sm:text-sm"
-                    />
-                  </div>
-                  <div className="flex items-center">
-                    <input
-                      type="checkbox"
-                      checked={app.redirectConfig?.openInNewTab !== false}
-                      onChange={e =>
-                        handleInputChange('redirectConfig', {
-                          ...app.redirectConfig,
-                          openInNewTab: e.target.checked
-                        })
-                      }
-                      className="h-4 w-4 text-indigo-600 focus:ring-indigo-500 border-gray-300 dark:border-gray-600 rounded"
-                    />
-                    <label className="ml-2 block text-sm text-gray-900 dark:text-gray-100">
-                      {t('admin.apps.edit.openInNewTab', 'Open in New Tab')}
-                    </label>
-                  </div>
-                  <div className="flex items-center">
-                    <input
-                      type="checkbox"
-                      checked={app.redirectConfig?.showWarning !== false}
-                      onChange={e =>
-                        handleInputChange('redirectConfig', {
-                          ...app.redirectConfig,
-                          showWarning: e.target.checked
-                        })
-                      }
-                      className="h-4 w-4 text-indigo-600 focus:ring-indigo-500 border-gray-300 dark:border-gray-600 rounded"
-                    />
-                    <label className="ml-2 block text-sm text-gray-900 dark:text-gray-100">
-                      {t('admin.apps.edit.showWarning', 'Show Warning Before Redirect')}
-                    </label>
-                  </div>
-                </div>
-              </div>
-            </div>
-          </div>
+          <RedirectConfigSection app={app} onChange={handleInputChange} t={t} />
         )}
 
         {/* Chat-specific sections - Only show for chat apps */}
